Ajuste na busca de Dieta Especial.
