+++ src/components/screens/DashboardGestaoProduto/index.jsx
@@ -3,7 +3,8 @@ import HTTP_STATUS from "http-status-codes";
 import CardStatusDeSolicitacao from "components/Shareable/CardStatusDeSolicitacao/CardStatusDeSolicitacao";
 import CardBody from "components/Shareable/CardBody";
 import { GESTAO_PRODUTO } from "configs/constants";
-import { formataCards, incluirDados, slugify } from "./helper";
+import { formataCards, incluirDados } from "./helper";
+import { slugify } from "../helper";
 import { dataAtual, deepCopy } from "helpers/utilities";
 import { listarCardsPermitidos } from "helpers/gestaoDeProdutos";
 import { getDashboardGestaoProdutos } from "services/produto.service";
