+++ src/components/screens/__tests__/helper.test.js
@@ -1,0 +1,9 @@
+import { slugify } from "../helper";
+
+describe("Teste slugfy", () => {
+  it("test slugfy strings", () => {
+    expect(slugify("Macarrão".toLocaleLowerCase())).toEqual(
+      slugify("macaRRao".toLocaleLowerCase())
+    );
+  });
+});
